Improves namespace :picture for user + views
Removes :picture column from User-table and replaces it with :avatar.
Also changes models to work with the new namespace and updates forms in
Devise.
As a bonus a user-picture is shown in the /dashboard when user is logged
in.

disclaimer:
Please take note that traces of :picture could be left in the app since
not all files were troughoutly checked.