Fix comments for !user_logged_in? and destroy
Added an if-statement in the deletebutton to check if user is signed in.
and added the first destroy scripts to make sure favorites and comments
are removed for the article that might get removed along the way. this
to ensure users are not confronted with unexisting pages and the server
doesn't  crash along this way.