Link routes to links and DB
Titles are now rendered with the @title string in each controller. If no
title is provided the system will render "Inspiratieplanet". Reviews are
now rendered in a json file (found in /reviews/reviews.json).
A minor bug has been removed in the review loop on the showpage.
the basic index page in public (later renamed to hoi.html) has been
removed since this will no longer be consulted in the running project.